requirements.txt
aiohttp==3.7.4.post0
aiohttp-socks==0.7.1
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
discord.py==1.7.3
emoji==1.7.0
fonttools==4.33.3
gTTS==2.2.4
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-socks==2.0.3
requests==2.27.1
six==1.16.0
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
wordcloud==1.8.1
yarl==1.7.2
